actions: Split out zerver.actions.create_user.
(cherry picked from commit cbad5739abcb7ab9878eb806eae806c681754165)
