+++ tools/test-api
@@ -31,13 +31,9 @@ with test_server_running(
 ):
     # zerver imports should happen after `django.setup()` is run
     # by the test_server_running decorator.
+    from zerver.actions.create_user import do_create_user, do_reactivate_user
     from zerver.actions.users import change_user_is_active
-    from zerver.lib.actions import (
-        do_create_user,
-        do_deactivate_realm,
-        do_reactivate_realm,
-        do_reactivate_user,
-    )
+    from zerver.lib.actions import do_deactivate_realm, do_reactivate_realm
     from zerver.lib.test_helpers import reset_emails_in_zulip_realm
     from zerver.lib.users import get_api_key
     from zerver.models import get_realm, get_user
