+++ zerver/management/commands/create_realm.py
@@ -3,7 +3,8 @@ from typing import Any
 
 from django.core.management.base import CommandError
 
-from zerver.lib.actions import do_create_realm, do_create_user
+from zerver.actions.create_user import do_create_user
+from zerver.lib.actions import do_create_realm
 from zerver.lib.management import ZulipBaseCommand
 from zerver.models import UserProfile
