+++ zerver/tests/test_decorators.py
@@ -13,6 +13,7 @@ from django.core.exceptions import ValidationError
 from django.http import HttpRequest, HttpResponse
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.create_user import do_reactivate_user
 from zerver.actions.users import change_user_is_active, do_deactivate_user
 from zerver.decorator import (
     authenticate_notify,
@@ -33,7 +34,6 @@ from zerver.lib.actions import (
     do_create_realm,
     do_deactivate_realm,
     do_reactivate_realm,
-    do_reactivate_user,
     do_set_realm_property,
 )
 from zerver.lib.cache import dict_to_items_tuple, ignore_unhashable_lru_cache, items_tuple_to_dict
