+++ zerver/tests/test_message_send.py
@@ -10,6 +10,7 @@ from django.http import HttpResponse
 from django.test import override_settings
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.create_user import do_create_user
 from zerver.actions.message_send import (
     build_message_send_dict,
     check_message,
@@ -27,12 +28,7 @@ from zerver.actions.message_send import (
 )
 from zerver.actions.streams import do_change_stream_post_policy
 from zerver.actions.users import do_change_can_forge_sender, do_deactivate_user
-from zerver.lib.actions import (
-    do_add_realm_domain,
-    do_create_realm,
-    do_create_user,
-    do_set_realm_property,
-)
+from zerver.lib.actions import do_add_realm_domain, do_create_realm, do_set_realm_property
 from zerver.lib.addressee import Addressee
 from zerver.lib.cache import cache_delete, get_stream_cache_key
 from zerver.lib.exceptions import JsonableError
