+++ zerver/tests/test_users.py
@@ -12,6 +12,7 @@ from django.test import override_settings
 from django.utils.timezone import now as timezone_now
 
 from confirmation.models import Confirmation
+from zerver.actions.create_user import do_create_user, do_reactivate_user
 from zerver.actions.invites import do_create_multiuse_invite_link, do_invite_users
 from zerver.actions.message_send import get_recipient_info
 from zerver.actions.users import (
@@ -21,12 +22,7 @@ from zerver.actions.users import (
     do_deactivate_user,
     do_delete_user,
 )
-from zerver.lib.actions import (
-    do_create_user,
-    do_mute_user,
-    do_reactivate_user,
-    do_set_realm_property,
-)
+from zerver.lib.actions import do_mute_user, do_set_realm_property
 from zerver.lib.avatar import avatar_url, get_gravatar_url
 from zerver.lib.bulk_create import create_users
 from zerver.lib.create_user import copy_default_settings
